+++ packages/react/src/components/PlaneStack3D/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2024
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+// Re-export everything from src/index.js
+export * from './src/index.js';
+export { default } from './src/index.js';