+++ packages/web-components/src/components/style-picker/__tests__/__snapshots__/style-picker.test.snap.js
@@ -1,8 +1,9 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
 
-snapshots["clabs-style-picker should render single variant color picker module"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant color picker module'
+] = `<clabs-style-picker
   align="bottom"
   kind="single"
   open=""
@@ -25,8 +26,9 @@ snapshots["clabs-style-picker should render single variant color picker module"]
 `;
 /* end snapshot clabs-style-picker should render single variant color picker module */
 
-snapshots["clabs-style-picker should render single variant icon picker module"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant icon picker module'
+] = `<clabs-style-picker
   align="bottom"
   kind="single"
   open=""
@@ -49,8 +51,9 @@ snapshots["clabs-style-picker should render single variant icon picker module"] 
 `;
 /* end snapshot clabs-style-picker should render single variant icon picker module */
 
-snapshots["clabs-style-picker should render single variant with color picker module"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant with color picker module'
+] = `<clabs-style-picker
   align="bottom"
   kind="single"
   open=""
@@ -73,8 +76,9 @@ snapshots["clabs-style-picker should render single variant with color picker mod
 `;
 /* end snapshot clabs-style-picker should render single variant with color picker module */
 
-snapshots["clabs-style-picker should render single variant with icon picker module"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant with icon picker module'
+] = `<clabs-style-picker
   align="bottom"
   kind="single"
   open=""
@@ -97,8 +101,9 @@ snapshots["clabs-style-picker should render single variant with icon picker modu
 `;
 /* end snapshot clabs-style-picker should render single variant with icon picker module */
 
-snapshots["clabs-style-picker should render single variant with pictogram picker module"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant with pictogram picker module'
+] = `<clabs-style-picker
   align="bottom"
   kind="single"
   open=""
@@ -121,8 +126,9 @@ snapshots["clabs-style-picker should render single variant with pictogram picker
 `;
 /* end snapshot clabs-style-picker should render single variant with pictogram picker module */
 
-snapshots["clabs-style-picker should render single variant with color picker options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant with color picker options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -2315,8 +2321,9 @@ snapshots["clabs-style-picker should render single variant with color picker opt
 `;
 /* end snapshot clabs-style-picker should render single variant with color picker options */
 
-snapshots["clabs-style-picker should render single variant with icon picker options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker should render single variant with icon picker options'
+] = `<clabs-style-picker
   align="bottom"
   heading="Choose icon"
   kind="single"
@@ -2393,8 +2400,8 @@ snapshots["clabs-style-picker should render single variant with icon picker opti
 `;
 /* end snapshot clabs-style-picker should render single variant with icon picker options */
 
-snapshots["clabs-style-picker should select a color option"] = 
-`<clabs-style-picker
+snapshots['clabs-style-picker should select a color option'] =
+  `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -4587,8 +4594,9 @@ snapshots["clabs-style-picker should select a color option"] =
 `;
 /* end snapshot clabs-style-picker should select a color option */
 
-snapshots["clabs-style-picker single variant should render single variant with color picker options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant should render single variant with color picker options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -6781,8 +6789,8 @@ snapshots["clabs-style-picker single variant should render single variant with c
 `;
 /* end snapshot clabs-style-picker single variant should render single variant with color picker options */
 
-snapshots["clabs-style-picker single variant should select a color option"] = 
-`<clabs-style-picker
+snapshots['clabs-style-picker single variant should select a color option'] =
+  `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -8975,8 +8983,9 @@ snapshots["clabs-style-picker single variant should select a color option"] =
 `;
 /* end snapshot clabs-style-picker single variant should select a color option */
 
-snapshots["clabs-style-picker single variant color picker should render single variant with color picker options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should render single variant with color picker options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -11169,8 +11178,9 @@ snapshots["clabs-style-picker single variant color picker should render single v
 `;
 /* end snapshot clabs-style-picker single variant color picker should render single variant with color picker options */
 
-snapshots["clabs-style-picker single variant color picker should select a color option"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should select a color option'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -11186,6 +11196,7 @@ snapshots["clabs-style-picker single variant color picker should select a color 
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -13363,8 +13374,9 @@ snapshots["clabs-style-picker single variant color picker should select a color 
 `;
 /* end snapshot clabs-style-picker single variant color picker should select a color option */
 
-snapshots["clabs-style-picker single variant icon picker should render single variant with icon picker options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant icon picker should render single variant with icon picker options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -13378,6 +13390,7 @@ snapshots["clabs-style-picker single variant icon picker should render single va
   <cds-icon-button
     align="top"
     close-on-activation=""
+    danger-description="danger"
     kind="ghost"
     size="md"
     slot="trigger"
@@ -13486,8 +13499,9 @@ snapshots["clabs-style-picker single variant icon picker should render single va
 `;
 /* end snapshot clabs-style-picker single variant icon picker should render single variant with icon picker options */
 
-snapshots["clabs-style-picker single variant icon picker should select a icon option"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant icon picker should select a icon option'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -13501,6 +13515,7 @@ snapshots["clabs-style-picker single variant icon picker should select a icon op
   <cds-icon-button
     align="top"
     close-on-activation=""
+    danger-description="danger"
     kind="ghost"
     size="md"
     slot="trigger"
@@ -13609,8 +13624,9 @@ snapshots["clabs-style-picker single variant icon picker should select a icon op
 `;
 /* end snapshot clabs-style-picker single variant icon picker should select a icon option */
 
-snapshots["clabs-style-picker flat variant color and icon picker should render color and icon picker flat variant"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should render color and icon picker flat variant'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -13913,8 +13929,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should render c
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should render color and icon picker flat variant */
 
-snapshots["clabs-style-picker single variant color picker should render single variant with color options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should render single variant with color options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -13930,6 +13947,7 @@ snapshots["clabs-style-picker single variant color picker should render single v
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -16107,8 +16125,9 @@ snapshots["clabs-style-picker single variant color picker should render single v
 `;
 /* end snapshot clabs-style-picker single variant color picker should render single variant with color options */
 
-snapshots["clabs-style-picker flat variant color and icon picker should render flat variant with color and icon options "] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should render flat variant with color and icon options '
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -16124,6 +16143,7 @@ snapshots["clabs-style-picker flat variant color and icon picker should render f
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -16411,8 +16431,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should render f
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should render flat variant with color and icon options  */
 
-snapshots["clabs-style-picker flat variant color and icon picker should select a different color and icon options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should select a different color and icon options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -16428,6 +16449,7 @@ snapshots["clabs-style-picker flat variant color and icon picker should select a
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -16715,8 +16737,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should select a
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should select a different color and icon options */
 
-snapshots["clabs-style-picker single variant color picker should enable search"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should enable search'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -18910,8 +18933,9 @@ snapshots["clabs-style-picker single variant color picker should enable search"]
 `;
 /* end snapshot clabs-style-picker single variant color picker should enable search */
 
-snapshots["clabs-style-picker flat variant color and icon picker should enable search"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should enable search'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -19215,8 +19239,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should enable s
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should enable search */
 
-snapshots["clabs-style-picker single variant color picker should enable search in single variant"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should enable search in single variant'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -19233,6 +19258,7 @@ snapshots["clabs-style-picker single variant color picker should enable search i
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -21410,8 +21436,9 @@ snapshots["clabs-style-picker single variant color picker should enable search i
 `;
 /* end snapshot clabs-style-picker single variant color picker should enable search in single variant */
 
-snapshots["clabs-style-picker flat variant color and icon picker should enable search in flat variant"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should enable search in flat variant'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -21428,6 +21455,7 @@ snapshots["clabs-style-picker flat variant color and icon picker should enable s
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -21715,8 +21743,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should enable s
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should enable search in flat variant */
 
-snapshots["clabs-style-picker flat variant color and icon picker should search color and icon"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker flat variant color and icon picker should search color and icon'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -21733,6 +21762,7 @@ snapshots["clabs-style-picker flat variant color and icon picker should search c
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -22035,8 +22065,9 @@ snapshots["clabs-style-picker flat variant color and icon picker should search c
 `;
 /* end snapshot clabs-style-picker flat variant color and icon picker should search color and icon */
 
-snapshots["clabs-style-picker disclosed variant color and icon picker should render disclosed variant with color and icon options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker disclosed variant color and icon picker should render disclosed variant with color and icon options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -22052,6 +22083,7 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should ren
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -24326,8 +24358,9 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should ren
 `;
 /* end snapshot clabs-style-picker disclosed variant color and icon picker should render disclosed variant with color and icon options */
 
-snapshots["clabs-style-picker disclosed variant color and icon picker should select a different color and icon options"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker disclosed variant color and icon picker should select a different color and icon options'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -24343,6 +24376,7 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should sel
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -26617,8 +26651,9 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should sel
 `;
 /* end snapshot clabs-style-picker disclosed variant color and icon picker should select a different color and icon options */
 
-snapshots["clabs-style-picker disclosed variant color and icon picker should enable search in disclosed variant"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker disclosed variant color and icon picker should enable search in disclosed variant'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -26635,6 +26670,7 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should ena
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -28909,8 +28945,9 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should ena
 `;
 /* end snapshot clabs-style-picker disclosed variant color and icon picker should enable search in disclosed variant */
 
-snapshots["clabs-style-picker disclosed variant color and icon picker should search color and icon"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker disclosed variant color and icon picker should search color and icon'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle="Try a different search"
   empty-state-title="No results found"
@@ -28927,6 +28964,7 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should sea
     aria-controls="style-picker"
     aria-expanded="true"
     close-on-activation=""
+    danger-description="danger"
     id="trigger"
     kind="ghost"
     size="md"
@@ -31259,8 +31297,9 @@ snapshots["clabs-style-picker disclosed variant color and icon picker should sea
 `;
 /* end snapshot clabs-style-picker disclosed variant color and icon picker should search color and icon */
 
-snapshots["clabs-style-picker single variant color picker should render loading state"] = 
-`<clabs-style-picker
+snapshots[
+  'clabs-style-picker single variant color picker should render loading state'
+] = `<clabs-style-picker
   align="bottom"
   empty-state-subtitle=""
   empty-state-title=""
@@ -33453,4 +33492,3 @@ snapshots["clabs-style-picker single variant color picker should render loading 
 </clabs-style-picker>
 `;
 /* end snapshot clabs-style-picker single variant color picker should render loading state */
-